IMCTF 2021/misc/pixcels-equal-cells/requirements.txt
numpy==1.21.3
pandas==1.3.5
Pillow==8.4.0
pylightxl==1.58
python-dateutil==2.8.2
pytz==2021.3
pyxlsb==1.0.9
six==1.16.0
